Fix: Add dtype conversion for DECODE_MULTI cache writes
The rotary_kv_update_sdpa_multi function was writing K/V tensors directly
to the cache without converting to cache dtype. This caused issues when
the model outputs bf16 but the cache is float32, similar to the PREFILL
and DECODE_SINGLE bugs fixed earlier.

+++ yalis/attention/sdpa_and_flex.py
@@ -476,8 +476,8 @@ def rotary_kv_update_sdpa_multi(
         q, k = roped_tensors
 
     index_kv = index_pos.view(B, 1, T, 1).expand(B, nh, T, hs)
-    k_cache[:B].scatter_(dim=2, index=index_kv, src=k)
-    v_cache[:B].scatter_(dim=2, index=index_kv, src=v)
+    k_cache[:B].scatter_(dim=2, index=index_kv, src=k.to(k_cache.dtype))
+    v_cache[:B].scatter_(dim=2, index=index_kv, src=v.to(v_cache.dtype))
 
     # Create the mask
     arange_t = torch.arange(k_cache.size(-2), device=k_cache.device).view(
